refactor: перенести часть логики храненилища жанров в хранилище фильмов
