+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/genre/GenreDao.java
@@ -3,27 +3,8 @@ package ru.yandex.practicum.filmorate.storage.dao.genre;
 import ru.yandex.practicum.filmorate.model.film.Genre;
 
 import java.util.Collection;
-import java.util.Set;
 
 public interface GenreDao {
-    /**
-     * Метод добавляет в хранилище связь между
-     * фильмом и жанром.
-     *
-     * @param filmID идентификатор фильма.
-     * @param genres идентификатор жанра.
-     */
-    void add(long filmID, Set<Genre> genres);
-
-    /**
-     * Метод обновляет связь в хранилище между
-     * фильмом и пользователем.
-     *
-     * @param filmID идентификатор фильма.
-     * @param genres идентификатор жанра.
-     */
-    void update(long filmID, Set<Genre> genres);
-
     /**
      * Меьтод возвращает жанр из хранилища
      * по его идентификатору.
@@ -41,16 +22,6 @@ public interface GenreDao {
      * хранилище.
      */
     Collection<Genre> getAll();
-
-    /**
-     * Метод возвращает все жанры фильма, по
-     * его идентификатору.
-     *
-     * @param filmID идентификатор фильма.
-     * @return Уникальная коллекция с жанрами,
-     * принадлежащая фильму.
-     */
-    Set<Genre> getGenres(long filmID);
 
     /**
      * Метод проверяет содержится ли в хранилище
