+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -8,21 +8,24 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.film.Film;
+import ru.yandex.practicum.filmorate.model.film.Genre;
 import ru.yandex.practicum.filmorate.model.film.Mpa;
-import ru.yandex.practicum.filmorate.storage.Storage;
+import ru.yandex.practicum.filmorate.storage.dao.genre.GenreDaoImpl;
 
 import java.sql.Date;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import static java.lang.String.format;
 import static ru.yandex.practicum.filmorate.service.Service.DEPENDENCY_MESSAGE;
 
 @Slf4j
 @Component("FilmDbStorage")
-public class FilmDbStorage implements Storage<Film> {
+public class FilmDbStorage implements FilmStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Autowired
@@ -109,6 +112,52 @@ public class FilmDbStorage implements Storage<Film> {
             log.trace("Не найден фильм ID_{}.", filmID);
             return false;
         }
+    }
+
+    @Override
+    public void addGenres(long filmID, Set<Genre> genres) {
+        log.debug("add({}, {})", filmID, genres);
+        for (Genre genre : genres) {
+            jdbcTemplate.update(""
+                    + "INSERT INTO film_genres (film_id, genre_id) "
+                    + "VALUES (?, ?)", filmID, genre.getId());
+            log.trace("Фильму ID_{} добавлен жанр ID_{}.", filmID, genre.getId());
+        }
+    }
+
+    @Override
+    public void updateGenres(long filmID, Set<Genre> genres) {
+        log.debug("update({}, {})", filmID, genres);
+        delete(filmID);
+        addGenres(filmID, genres);
+    }
+
+    @Override
+    public Set<Genre> getGenres(long filmID) {
+        log.debug("getGenres({}).", filmID);
+        Set<Genre> genres = new HashSet<>(jdbcTemplate.query(format(""
+                + "SELECT f.genre_id, g.name "
+                + "FROM film_genres AS f "
+                + "LEFT OUTER JOIN genres AS g ON f.genre_id = g.genre_id "
+                + "WHERE f.film_id=%d "
+                + "ORDER BY g.genre_id", filmID), new GenreDaoImpl.GenreMapper()));
+        log.trace("Возвращены все жанры для фильма ID_{}: {}.", filmID, genres);
+        return genres;
+    }
+
+    /**
+     * Метод удаляет все связи фильма с
+     * жанрами.
+     *
+     * @param filmID идентификатор фильма.
+     */
+    private void delete(long filmID) {
+        log.debug("delete({}).", filmID);
+        jdbcTemplate.update(""
+                + "DELETE "
+                + "FROM film_genres "
+                + "WHERE film_id=?", filmID);
+        log.debug("Удалены все жанры у фильма {}.", filmID);
     }
 
     private static class FilmMapper implements RowMapper<Film> {
